Migrate Dropdown to TypeScript
