+++ package.json
@@ -75,6 +75,7 @@
     "@types/lodash": "^4.14.168",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
+    "@types/react-onclickoutside": "^6.7.3",
     "@types/redux-form": "^7.0.4",
     "@types/styled-components": "^5.1.7",
     "@typescript-eslint/eslint-plugin": "^4.13.0",
